winter_math::fft

Function serial_fft

pub fn serial_fft<B, E>(values: &mut [E], twiddles: &[B])
where
    B: StarkField,
    E: FieldElement<BaseField = B>,

Executes a single-threaded version of the FFT algorithm on the provided values.

The evaluation is done in-place, meaning the function does not allocate any additional memory, and the results are written back into values.

The twiddles needed for evaluation can be obtained via fft::get_twiddles() function using values.len() as the domain size parameter. This implies that twiddles.len() must be equal to values.len() / 2.

Panics

Panics if:

Length of values is not a power of two.

Length of twiddles is not values.len() / 2.

Field specified by B does not contain a multiplicative subgroup of size values.len().