winter_math::fields::f128

Struct BaseElement

pub struct BaseElement(/* private fields */);

Represents a base field element.

Internal values are stored in their canonical form in the range [0, M). The backing type is u128.

Implementations

impl BaseElement

pub const fn new(value: u128) -> Self

Creates a new field element from a u128 value. If the value is greater than or equal to the field modulus, modular reduction is silently performed. This function can also be used to initialize constants.