winter_math

Trait ExtensionOf

pub trait ExtensionOf<E: FieldElement>: From<E> {
    // Required method
    fn mul_base(self, other: E) -> Self;
}

Specifies that a field is an extension of another field.

Currently, this implies the following:

An element in the base field can be converted into an element in the extension field.

An element in the extension field can be multiplied by a base field element directly. This can be used for optimization purposes as such multiplication could be much more efficient than multiplication of two extension field elements.

Required Methods

fn mul_base(self, other: E) -> Self

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.